fix: Update stories to prevent duplicate kind warning
