--- modules/_labs/color-picker/react/stories/stories_testing.tsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import {ColorPicker} from '@workday/canvas-kit-labs-react-color-picker';
-
-export default {
-  title: 'Testing/React/Labs/Color Picker',
-  component: ColorPicker,
-};
-
-export const InputInteraction = () => {
-  const textAreaRef = React.useRef(null);
-  const [color, setColor] = React.useState('');
-  const handleColorChange = (c: string) => {
-    setColor(c);
-    textAreaRef.current?.focus();
-  };
-  return (
-    <div className="App">
-      <ColorPicker showCustomHexInput onColorChange={handleColorChange} />
-      <label htmlFor="test">Text Area</label>
-      <textarea id="test" style={{color: color}} ref={textAreaRef} />
-    </div>
-  );
-};
