+++ modules/button/react/stories/visual-testing/theming/stories_Button.tsx
@@ -4,7 +4,7 @@ import {Button} from '../../../';
 import {ButtonStates} from '../stories_Button';
 
 export default withSnapshotsEnabled({
-  title: 'Testing/React/Buttons/Button/Button',
+  title: 'Testing/React/Buttons/Button/Button/Theming',
   component: Button,
   parameters: {
     canvasProviderDecorator: {
