+++ modules/button/react/stories/visual-testing/theming/stories_DeleteButton.tsx
@@ -4,7 +4,7 @@ import {DeleteButton} from '../../../';
 import {DeleteButtonStates} from '../stories_DeleteButton';
 
 export default withSnapshotsEnabled({
-  title: 'Testing/React/Buttons/Button/Delete Button',
+  title: 'Testing/React/Buttons/Button/Delete Button/Theming',
   component: DeleteButton,
   parameters: {
     canvasProviderDecorator: {
