+++ modules/button/react/stories/visual-testing/theming/stories_ToolbarDropdownButton.tsx
@@ -4,7 +4,7 @@ import {ToolbarDropdownButton} from '../../../';
 import {ToolbarDropdownButtonStates} from '../stories_ToolbarDropdownButton';
 
 export default withSnapshotsEnabled({
-  title: 'Testing/React/Buttons/Button/Toolbar Dropdown Button',
+  title: 'Testing/React/Buttons/Button/Toolbar Dropdown Button/Theming',
   component: ToolbarDropdownButton,
   parameters: {
     canvasProviderDecorator: {
