+++ modules/button/react/stories/visual-testing/theming/stories_ToolbarIconButton.tsx
@@ -4,7 +4,7 @@ import {ToolbarIconButton} from '../../../';
 import {ToolbarIconButtonStates} from '../stories_ToolbarIconButton';
 
 export default withSnapshotsEnabled({
-  title: 'Testing/React/Buttons/Button/Toolbar Icon Button',
+  title: 'Testing/React/Buttons/Button/Toolbar Icon Button/Theming',
   component: ToolbarIconButton,
   parameters: {
     canvasProviderDecorator: {
