+++ modules/form-field/react/stories/visual-testing/theming/stories_ColorInput.tsx
@@ -4,7 +4,7 @@ import {ColorInput} from '../../../../../color-picker/react';
 import {ColorInputStates} from '../stories_ColorInput';
 
 export default withSnapshotsEnabled({
-  title: 'Testing/React/Inputs/Color Picker/Color Input',
+  title: 'Testing/React/Inputs/Color Picker/Color Input/Theming',
   component: ColorInput,
   parameters: {
     canvasProviderDecorator: {
