+++ modules/form-field/react/stories/visual-testing/theming/stories_Radio.tsx
@@ -4,7 +4,7 @@ import {Radio} from '../../../../../radio/react';
 import {RadioStates} from '../stories_Radio';
 
 export default withSnapshotsEnabled({
-  title: 'Testing/React/Inputs/Radio',
+  title: 'Testing/React/Inputs/Radio/Theming',
   component: Radio,
   parameters: {
     canvasProviderDecorator: {
