+++ modules/form-field/react/stories/visual-testing/theming/stories_Select.tsx
@@ -4,7 +4,7 @@ import {Select} from '../../../../../select/react';
 import {SelectStates} from '../stories_Select';
 
 export default withSnapshotsEnabled({
-  title: 'Testing/React/Inputs/Select',
+  title: 'Testing/React/Inputs/Select/Theming',
   component: Select,
   parameters: {
     canvasProviderDecorator: {
