+++ modules/form-field/react/stories/visual-testing/theming/stories_TextArea.tsx
@@ -4,7 +4,7 @@ import {TextArea} from '../../../../../text-area/react';
 import {TextAreaStates} from '../stories_TextArea';
 
 export default withSnapshotsEnabled({
-  title: 'Testing/React/Inputs/TextArea',
+  title: 'Testing/React/Inputs/TextArea/Theming',
   component: TextArea,
   parameters: {
     canvasProviderDecorator: {
