+++ modules/form-field/react/stories/visual-testing/theming/stories_TextInput.tsx
@@ -4,7 +4,7 @@ import {TextInput} from '../../../../../text-input/react';
 import {TextInputStates} from '../stories_TextInput';
 
 export default withSnapshotsEnabled({
-  title: 'Testing/React/Inputs/Text Input',
+  title: 'Testing/React/Inputs/Text Input/Theming',
   component: TextInput,
   parameters: {
     canvasProviderDecorator: {
