+++ modules/popup/react/stories/stories_testing.tsx
@@ -9,6 +9,8 @@ export default {
   component: Popup,
 };
 
+export * from './stories_VisualTesting';
+
 export const MultiplePopups = () => {
   const popup1 = usePopup();
   const popup2 = usePopup();
